+++ Game/Game/GameLogic.h
@@ -9,68 +9,13 @@
 
 #include "GameGraphic.h"
 #include "Header.h"
-
-#pragma region GAME CONFIG
-
-#define PLAYER_H 50
-#define PLAYER_W 50
-
-#pragma endregion
+#include "GameObject.h"
+#include "Player.h"
+#include "GameObstacles.h"
 
 using namespace std;
 using namespace std::this_thread;
 using namespace std::chrono;
-
-#pragma region GameObject
-
-class GameObject {
-private:
-protected:
-	int x, y; // coordinate of bottom-left corner
-	int w, h; // box size
-	int row;  // path number for checking collision
-public:
-	GameObject() : x(0), y(0), w(0), h(0), row(0) {}
-	GameObject(int x, int y, int w, int h, int row) : x(x), y(y), w(w), h(h), row(0) {}
-
-	void move(int x, int y)
-	{
-		this->x += x;
-		this->y += y;
-	}
-	virtual void render() = 0;
-	//virtual void collision() = 0;
-};
-
-class Obstacle : public GameObject {
-private:
-protected:
-	int veclocity;
-	DIRECTION direction; // may need a direction
-public:
-	Obstacle() : GameObject(), veclocity(0), direction(DIRECTION::RIGHT) {}
-	Obstacle(int x, int y, int w, int h, int row, int veclocity, DIRECTION direction = DIRECTION::RIGHT)
-		: GameObject(x, y, w, h, row), veclocity(veclocity), direction(direction) {}
-
-	void render() = 0;
-	//void collision();
-};
-
-#pragma region Obstacles
-class Something : public Obstacle {};
-class Will : public Obstacle {};
-class BeDefineLater : public Obstacle {};
-#pragma endregion
-
-class Player : public GameObject {
-private:
-protected:
-public:
-	Player() : GameObject() {}
-	Player(int x, int y, int w = PLAYER_W, int h = PLAYER_H) : GameObject(x, y, w, h, 0) {}
-	void render();
-};
-#pragma endregion
 
 #pragma region GameCore
 
@@ -79,9 +24,9 @@ private:
 	GAMESTATE state;
 	GameGraphic gph;
 	int level;
-	int currentInput;
 	//will need more param
 public:
+	GameCore(GAMESTATE state = GAMESTATE::PLAYING);
 	void start()
 	{
 		thread t1(&GameCore::draw, this);
@@ -126,6 +71,8 @@ public:
 		if (bKey[3] == 1) {
 			player->move(1, 0);
 		}
+
+		delete[] bKey;
 	}
 
 	void gameLogic()
@@ -147,6 +94,9 @@ public:
 			//if pass -> send signal to check
 			//else set flag game over
 		}
+
+		delete player;
+		//delete obstacles too
 	}
 
 	void initObstacles(int level, vector<vector<GameObject*>> & obs)
